Separate Avatar as a component.
