+++ src/app/core/common/Navbar.js
@@ -4,9 +4,9 @@ import { withStyles } from '@material-ui/core/styles'
 import { withRouter } from 'react-router'
 import { rootPath } from '../globals'
 import classNames from 'classnames'
+import Avatar from './Avatar'
 import {
   AppBar,
-  Avatar,
   Button,
   Divider,
   Drawer,
@@ -23,7 +23,6 @@ import {
 import ChevronLeftIcon from '@material-ui/icons/ChevronLeft'
 import SearchIcon from '@material-ui/icons/Search'
 import MenuIcon from '@material-ui/icons/Menu'
-import { LOCAL_STORAGE_NAMESPACE } from './pf9-storage'
 
 const drawerWidth = 240
 
@@ -116,20 +115,9 @@ const styles = theme => ({
     display: 'flex',
     alignItems: 'center'
   },
-  avatar: {
-    position: 'relative',
-    float: 'right'
-  },
   selector: {
     position: 'relative',
     float: 'right'
-  },
-  avatarImg: {
-    backgroundColor: theme.palette.primary.dark,
-    marginRight: theme.spacing.unit,
-    fontSize: theme.spacing.unit * 2,
-    height: theme.spacing.unit * 3,
-    width: theme.spacing.unit * 3
   },
   menuSearch: {
     outline: 'none',
@@ -215,39 +203,6 @@ class Navbar extends React.Component {
     </div>
   }
 
-  renderAvatar = () => {
-    const userName = JSON.parse(localStorage.getItem(LOCAL_STORAGE_NAMESPACE)).username || ''
-    const { classes } = this.props
-    const { anchorEl } = this.state
-
-    return <div className={classes.avatar}>
-      <Button
-        aria-owns={anchorEl ? 'user-menu' : null}
-        aria-haspopup="true"
-        onClick={this.handleClick('anchorEl')}
-        color="inherit"
-        disableRipple
-      >
-        <Avatar className={classes.avatarImg}>{userName.charAt(0)}</Avatar>
-        <Typography color="inherit" variant="body1">
-          {userName}
-        </Typography>
-      </Button>
-      <Menu
-        id="user-menu"
-        anchorEl={anchorEl}
-        open={Boolean(anchorEl)}
-        onClose={this.handleClose('anchorEl')}
-        getContentAnchorEl={null}
-        anchorOrigin={{vertical: 'bottom', horizontal: 'left'}}
-      >
-        <MenuItem onClick={this.handleClose('anchorEl')}>Change Password</MenuItem>
-        <MenuItem onClick={this.handleClose('anchorEl')}>SSH Keys</MenuItem>
-        <MenuItem onClick={this.navTo('/ui/logout')}>Sign Out</MenuItem>
-      </Menu>
-    </div>
-  }
-
   render () {
     const { classes, links } = this.props
     const { open } = this.state
@@ -294,7 +249,7 @@ class Navbar extends React.Component {
               <div className={classes.rightTools}>
                 {this.renderSelector('Region', [`AWS-US-West-1-Test`, `KVM-Neutron`])}
                 {this.renderSelector('Tenant', [`Dev Team Tenant`, `Test Tenant`])}
-                {this.renderAvatar()}
+                <Avatar />
               </div>
             </Toolbar>
           </AppBar>
